RGBA bugfix

+++ blades/stm32l4_ws2811.h
@@ -172,7 +172,7 @@ public:
   }
 
   void BeginFrame() {
-    while (Color8::num_bytes(byteorder_) * num_leds_ > sizeof(displayMemory)) {
+    while (Color8::num_bytes(byteorder_) * num_leds_ > (int)sizeof(displayMemory)) {
       STDOUT.print("Display memory is not big enough, increase maxLedsPerStrip!");
       num_leds_ /= 2;
     }
